requirements.txt
matplotlib
numpy
Pillow
tqdm
torch
torchvision